requirements.txt
langchain
langchain-groq
langchain-core
langsmith
python-dotenv